tasks/views.py
from rest_framework.views import APIView
from rest_framework.response import Response
from .models import Task
from .serializers import TaskSerializer


class TaskListAPIView(APIView):
    def get(self, request)