pyproject.toml
[project]
name = "drf-study"
version = "0.1.0"
description = ""
authors = [
    {name = "Tolik_mac",email = "ttserunyan@inbox.ru"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
